Move milkman functions into core egp-16 file
