+++ egps/egp-16/test.ts
@@ -3,20 +3,12 @@ import { loadFixture } from '@nomicfoundation/hardhat-network-helpers';
 import { network, ethers } from 'hardhat'
 
 // Artifacts
-import iYearnVault from '../../elf-contracts/artifacts/contracts/interfaces/IYearnVault.sol/IYearnVault.json'
 import iERC20 from '../../elf-contracts/artifacts/contracts/interfaces/IERC20.sol/IERC20.json'
 import iCurvePool from '../../artifacts/egps/egp-16/CurvelpPriceChecker.sol/iCurvePool.json'
 import iPriceChecker from '../../artifacts/egps/egp-16/CurvelpPriceChecker.sol/IPriceChecker.json'
 import iCurveMetaPool from '../../artifacts/egps/egp-16/CurvelpPriceChecker.sol/iCurveMetaPool.json'
-import iMilkman from '../../artifacts/egps/egp-16/CurvelpPriceChecker.sol/iMilkman.json'
 
-import { yearnPools, crvPools, TREASURY_ADDRESS, withdrawFromYearn, deployPriceChecker } from "./egp-16";
-
-// Helpers
-import * as addresses from '../helpers/addresses'
-import { createCallHash } from '../helpers/hashing'
-import { CurvelpPriceChecker__factory } from '../../typechain'
-import { BigNumber } from "ethers";
+import { yearnPools, crvPools, TREASURY_ADDRESS, withdrawFromYearn, deployPriceChecker, swapviaMilkman } from "./egp-16";
 
 // Run the test on a mainnet fork
 describe("Run unwinding part 1, mainnet fork", function() {
@@ -227,50 +219,7 @@ describe("Run unwinding part 1, mainnet fork", function() {
     it("sends swap events to milkman", async function() {
         const { signer, address } = await loadFixture(priceCheckerDeployFixture);
 
-        const iMilkmanContract = new ethers.Contract(
-            '0x11C76AD590ABDFFCD980afEC9ad951B160F02797',
-            iMilkman.abi,
-            signer
-        );
-
-        for (let i in crvPools) {
-            const priceCheckerData = ethers.utils.defaultAbiCoder.encode(
-                ["uint256", "bool", "uint256", "int128", "address"],
-                // 1% slippage param
-                [100, crvPools[i].isInt128, crvPools[i].i, crvPools[i].i, crvPools[i].pool]
-            );
-
-            const lpTokenContract = new ethers.Contract(
-                yearnPools[i].withdrawn,
-                iERC20.abi,
-                signer
-            );
-            const lpBalance = await lpTokenContract.balanceOf(TREASURY_ADDRESS);
-            const lpSymbol = await lpTokenContract.symbol();
-
-            const withdrawnTokenContract = new ethers.Contract(
-                crvPools[i].withdrawn,
-                iERC20.abi,
-                signer
-            );
-            const withdrawnSymbol = await withdrawnTokenContract.symbol();
-
-            console.log(
-                `Requesting an on-chain Swap via Milkman at 1% slippage.\n`,
-                `${lpBalance} ${lpSymbol} -> ${withdrawnSymbol}`,
-            )
-
-            const calledSwap = await iMilkmanContract.requestSwapExactTokensForTokens(
-                lpBalance,
-                yearnPools[i].withdrawn,
-                crvPools[i].withdrawn,
-                TREASURY_ADDRESS,
-                address,
-                priceCheckerData
-            );
-            console.log(calledSwap);
-        }
-
+        await swapviaMilkman(signer, address);
         console.log("No Reverts Returned.")
     })
 })
